Add a function to get raster averages for nodes
A new function is added to calculate mean values at the selected nodes
from a geo-referenced raster data. The function supprots binned mapping,
which is designed mainly to process NDVI values. See the code comments
for more information.
This commit changed the way the polygon attirbute is evaluated, and the
evaluation is mostly vectorized. This may break some backward
compatibility, but some magics are added to maintain backward
compability.